Added HTTP server that stores and serves files. Delegates storage operations to Storage struct.
HTTP server accepts PUT, GET and DELETE requests with filename provided in URL.
The binary starts the server using host and data-dir provided as command line parameters.
In case of server errors the errors are logged and additionally returned as JSON message.

Added integration tests that run the binary and then issue HTTP calls. Checks
PUT, GET and DELETE for a single file. Also checks that uploading the same file
multiple time using different names does not lead to duplicates on the file system.
The implementation will check the hash of the file and only store unique files.
This will be implemented in Storage struct.

Storage struct is a placeholder that does not do anything at the moment. But the
API should be able to support multiple backends. We pass context.Context everywhere to
allow cancellation and timeouts in case we decide to store files on a remote
service, such as Amazon S3 later.

At the moment, the integration tests fail, because the Storage is not implemented yet.