Added sqlite3 third party dependency, which will be used to store metadata.
We need to store metadata for the files, such as file name, id, hash. It
would be possible to store it as bare files on the filesystem, but because
we require getting files by id, name and also checking how many files have
the same hash, using an sqlite database saves noticeable amount of code.

Will implement the storage using sqlite database next. Vendoring the database
and checking it in git, to allow building immediately after checkout without
and extra steps.