Store the files on disk and metadata in sqlite. Working HTTP file storage and retrieval.
Implemented the Storage struct. Storing metadata in sqlite and files on disk.
The HTTP service now works and allows uploading, downloading and deleting files.
This is checked in integration_test.go, all tests pass.

Storage layer create an sqlite database on the first run with one table
that contains file metadata (id, name sha256).

When file is uploaded, the contents is stored in one directory using sha256
of the content as name. The metadata is stored in sqlite.

When the file is uploaded, the Storage checks the sha256 of the uploaded file
to avoid storing the duplicated content twice.

The Storage also allows getting the metadata by id and by name. It provides
a method for getting the content and for deleting files.

When files are deleted, the service check how many files are using the same hash
to only delete the contents when it was the last file with the same hash.